requirements.txt
Flask==1.1.1
pandas
msoffcrypto-tool
openpyxl
azure.storage.blob
python-dateutil